move deriveTxType out of components folder
