+++ src/features/Dashboard/components/index.ts
@@ -5,5 +5,3 @@ export { default as TransactionLabel } from './TransactionLabel';
 export { WalletBreakdown } from './WalletBreakdown';
 export { default as NoTransactions } from './NoTransactions';
 export { default as MembershipPanel } from './MembershipPanel';
-export { deriveTxType } from './helpers';
-export { ITxHistoryType } from './types';
