--- src/features/Dashboard/components/types.ts
@@ -1,14 +1,0 @@
-import { ITxType } from '@types';
-
-enum IStandardTxType {
-  TRANSFER = 'TRANSFER',
-  OUTBOUND = 'OUTBOUND',
-  INBOUND = 'INBOUND'
-}
-
-export const ITxHistoryType = { ...ITxType, ...IStandardTxType };
-
-export type ITxHistoryType = Exclude<
-  Exclude<ITxType | IStandardTxType, ITxType.STANDARD>,
-  ITxType.UNKNOWN
->;
