+++ src/features/Dashboard/index.ts
@@ -1,1 +1,2 @@
 export { default as Dashboard } from './Dashboard';
+export { deriveTxType } from './helpers';
