+++ src/features/Dashboard/types.ts
@@ -1,4 +1,4 @@
-import { TURL } from '@types';
+import { TURL, ITxType } from '@types';
 export interface Action {
   icon: string;
   faded?: boolean;
@@ -6,3 +6,16 @@ export interface Action {
   description: string;
   link: string | TURL;
 }
+
+enum IStandardTxType {
+  TRANSFER = 'TRANSFER',
+  OUTBOUND = 'OUTBOUND',
+  INBOUND = 'INBOUND'
+}
+
+export const ITxHistoryType = { ...ITxType, ...IStandardTxType };
+
+export type ITxHistoryType = Exclude<
+  Exclude<ITxType | IStandardTxType, ITxType.STANDARD>,
+  ITxType.UNKNOWN
+>;
